Added gpx reading (still need to calculate distance from lat/long

+++ gps_reader/models.py
@@ -3,8 +3,8 @@ from django.db import models
 class Activity(models.Model):
     activity = models.CharField(max_length=200)
     sport = models.CharField(max_length=100)
-    time_s = models.IntegerField('time in seconds')
-    distance_m = models.IntegerField('distance in meters')
+    time_s = models.FloatField('time in seconds')
+    distance_m = models.FloatField('distance in meters')
     date = models.DateTimeField('start date/time')
 
     def __unicode__(self):
